refactor(grid): rename module for consistency
Rename module for consistency, update imports and references, and update
documentation.
