+++ src/goesdr/grid/grid_goesdr.py
